pyproject.toml
[project]
name = "mira"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mira-sdk>=0.1.14",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
]
